+++ src/Identity.swift
@@ -15,7 +15,7 @@ class Identity {
     init() {}
 
     // MARK: - Tracks the current user's account info
-    private(set) var account: Account? {
+    private(set) var account: AuthenticatedAccount? {
         didSet {
             print("ACCOUNT: INFO: Current user did change to:", account as Any)
             DispatchQueue.main.async {
@@ -24,11 +24,16 @@ class Identity {
         }
     }
 
-    func update(account: Account) {
+    /// Selected persona GUID.
+    private(set) var persona: String?
+
+    func update(account: AuthenticatedAccount, persona: String? = nil) {
+        self.persona = persona ?? account.distributors.first?.guid ?? "default"
         self.account = account
     }
 
     func logOut() {
+        persona = nil
         account = nil
     }
 }