+++ src/TenCenturiesPostRepository.swift
@@ -401,10 +401,11 @@ class TenCenturiesPostRepository: PostRepository, TenCenturiesService {
 
 
     // MARK: - Takes sundry other actions
-    func toggleStarred(post: Post, completion: @escaping (Result<[Post]>) -> Void) {
-        let url = URL(string: "/content/star/\(post.id)", relativeTo: TenCenturies.baseURL)!
+    func toggleStarred(post: Post, by persona: String, completion: @escaping (Result<[Post]>) -> Void) {
+        let url = URL(string: "/api/posts/star", relativeTo: TenCenturies.baseURL)!
         var request = URLRequest(url: url)
-        request.httpMethod = "POST"
+        request.httpMethod = post.you.starred ? "DELETE" : "POST"
+        request.httpBody = try! JSONEncoder().encode(["persona_guid": persona, "guid": post.id])
         let _ = send(request: request) { (result) in
             do {
                 let wrapper = try result.unwrap()
@@ -417,7 +418,7 @@ class TenCenturiesPostRepository: PostRepository, TenCenturiesService {
     }
 
     func pin(post: Post, color: Post.PinColor?, completion: @escaping (Result<[Post]>) -> Void) {
-        let url = URL(string: "/content/pin/\(post.id)", relativeTo: TenCenturies.baseURL)!
+        let url = URL(string: "/api/pin/\(post.id)", relativeTo: TenCenturies.baseURL)!
         var request = URLRequest(url: url)
         if let pin = color {
             request.httpMethod = "POST"
@@ -439,6 +440,7 @@ class TenCenturiesPostRepository: PostRepository, TenCenturiesService {
     }
 
     func repost(post: Post, completion: @escaping (Result<[Post]>) -> Void) {
+        // TODO: 10Cv5: repost should probably be handled as posting a quotation? maybe?
         let url = URL(string: "/content/repost/\(post.id)", relativeTo: TenCenturies.baseURL)!
         var request = URLRequest(url: url)
         request.httpMethod = "POST"